Rename project

+++ Coronavirus Herd Immunity/CovidApp/AppDelegate.swift
@@ -138,7 +138,7 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         let bundleID = Bundle.main.bundleIdentifier;
         print("Bundle ID: \(token) \(bundleID)");
         if let idDevice = StorageManager.shared.getIdentifierDevice(){
-            ApiManager.shared.setPushNotificationId(deviceId: idDevice, notificationId: token)
+            ApiManager.shared.setPushNotificationId(deviceId: Int64(idDevice), notificationId: token)
         }
         StorageManager.shared.setPushId(token)
         // 3. Save the token to local storeage and post to app server to generate Push Notification. ...
